+++ cli/tests/test_git_push.rs
@@ -154,7 +154,7 @@ fn test_git_push_current_bookmark() {
     let output = work_dir.run_jj(["git", "push"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=origin)..@
+    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=remote)..@
     Nothing changed.
     [EOF]
     ");
@@ -200,7 +200,7 @@ fn test_git_push_no_matching_bookmark() {
     let output = work_dir.run_jj(["git", "push"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=origin)..@
+    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=remote)..@
     Nothing changed.
     [EOF]
     ");
@@ -215,7 +215,7 @@ fn test_git_push_matching_bookmark_unchanged() {
     let output = work_dir.run_jj(["git", "push"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=origin)..@
+    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=remote)..@
     Nothing changed.
     [EOF]
     ");
@@ -261,7 +261,7 @@ fn test_git_push_other_remote_has_bookmark() {
     let output = work_dir.run_jj(["git", "push"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=origin)..@
+    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=remote)..@
     Nothing changed.
     [EOF]
     ");
@@ -2242,6 +2242,93 @@ fn test_git_push_rejected_by_remote() {
         [exit status: 1]
         ");
     });
+}
+
+#[test]
+fn test_git_push_custom_revset() {
+    let test_env = TestEnvironment::default();
+    set_up(&test_env);
+    let work_dir = test_env.work_dir("local");
+    // add a custom revset which simulates keeping a custom set of local only
+    // bookmarks.
+    test_env.add_config(
+        r#"
+    [revsets]
+    'git-push' = "remote_bookmarks(remote=remote)..@ ~untracked_remote_bookmarks(glob:'local/*')"
+    "#,
+    );
+    work_dir
+        .run_jj(["new", "bookmark2", "-m", "commit to be pushed"])
+        .success();
+    work_dir.run_jj(["new", "-m", "wip: stuff"]).success();
+    work_dir
+        .run_jj(["bookmark", "set", "local/stuff", "-r@"])
+        .success();
+    work_dir
+        .run_jj(["new", "-m", "commit which should pushed"])
+        .success();
+    //
+    let output = work_dir.run_jj(["log"]);
+    insta::assert_snapshot!(output, @r"
+    @  kpqxywon test.user@example.com 2001-02-03 08:05:17 048a6554
+    │  (empty) commit which should pushed
+    ○  yostqsxw test.user@example.com 2001-02-03 08:05:15 local/stuff 0712d559
+    │  (empty) wip: stuff
+    ○  vruxwmqv test.user@example.com 2001-02-03 08:05:14 a615282a
+    │  (empty) commit to be pushed
+    ○  zsuskuln test.user@example.com 2001-02-03 08:05:10 bookmark2 38a20473
+    │  (empty) description 2
+    │ ○  qpvuntsm test.user@example.com 2001-02-03 08:05:08 bookmark1 9b2e76de
+    ├─╯  (empty) description 1
+    ◆  zzzzzzzz root() 00000000
+    [EOF]
+    ");
+    // We should try to push everything except the commit on the "local/" bookmark.
+    let output = work_dir.run_jj(["git", "push"]);
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    Warning: Refusing to create new remote bookmark local/stuff@origin
+    Hint: Use --allow-new to push new bookmark. Use --remote to specify the remote to push to.
+    Nothing changed.
+    [EOF]
+    ");
+}
+
+// This tests that `exact: <remote>` isn't changed to a substring pattern.
+#[test]
+fn test_git_push_unambigous_remote() {
+    let test_env = TestEnvironment::default();
+    set_up(&test_env);
+    let work_dir = test_env.work_dir("local");
+    work_dir.run_jj(["new", "bookmark1"]).success();
+    let add_new_git_remote = |remote: &str| {
+        let remote_interpolated = format!("{remote}");
+        test_env
+            .run_jj_in(".", &["git", "init", &remote_interpolated])
+            .success();
+        let remote_dir = work_dir.dir(&remote_interpolated);
+        let remote_git_dir = git_repo_dir_for_jj_repo(&remote_dir);
+        work_dir
+            .run_jj([
+                "git",
+                "remote",
+                "add",
+                &remote_interpolated,
+                remote_git_dir.to_str().unwrap(),
+            ])
+            .success();
+    };
+    // Add some new remotes which could get matched on substrings.
+    add_new_git_remote("upstreeam");
+    add_new_git_remote("upstream");
+    work_dir.run_jj(["new", "bookmark1"]).success();
+    work_dir.run_jj(["desc", "-m", "Foo"]).success();
+    work_dir
+        .run_jj(["bookmark", "move", "bookmark1", "-r=@"])
+        .success();
+    let output = work_dir.run_jj(&["git", "push", "--remote", "upstream"]);
+    insta::assert_snapshot!(output, @r"
+        ");
 }
 
 #[must_use]